+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ### Changed
 - [android]Move DOM and render to seperate Action Object
+- [android]Refactor list-component sticky, use `List` instead of `Stack` to store sticky information.